requirements.txt
python-dotenv==1.0.1
selenium==4.18.1
Flask>=2.2.5
requests>=2.31.0
